refactoring pear and pear_channel
